+++ packages/library-base/src/commonMain/kotlin/io/realm/kotlin/internal/query/ObjectQuery.kt
@@ -32,10 +32,10 @@ import io.realm.kotlin.internal.interop.RealmCoreIndexOutOfBoundsException
 import io.realm.kotlin.internal.interop.RealmCoreInvalidQueryException
 import io.realm.kotlin.internal.interop.RealmCoreInvalidQueryStringException
 import io.realm.kotlin.internal.interop.RealmInterop
-import io.realm.kotlin.internal.interop.RealmPointer
 import io.realm.kotlin.internal.interop.RealmQueryPointer
 import io.realm.kotlin.internal.interop.RealmResultsPointer
 import io.realm.kotlin.internal.interop.inputScope
+import io.realm.kotlin.internal.schema.ClassMetadata
 import io.realm.kotlin.notifications.ResultsChange
 import io.realm.kotlin.query.RealmQuery
 import io.realm.kotlin.query.RealmResults
@@ -60,6 +60,8 @@ internal class ObjectQuery<E : BaseRealmObject> constructor(
         RealmInterop.realm_query_find_all(queryPointer)
     }
 
+    private val classMetadata: ClassMetadata? = realmReference.schemaMetadata[clazz.simpleName!!]
+
     internal constructor(
         realmReference: RealmReference,
         key: ClassKey,
@@ -72,12 +74,7 @@ internal class ObjectQuery<E : BaseRealmObject> constructor(
         key,
         clazz,
         mediator,
-        parseQuery(
-            realmReference.dbPointer,
-            key,
-            filter,
-            args
-        ),
+        parseQuery(realmReference, key, filter, args),
     )
 
     internal constructor(
@@ -148,7 +145,7 @@ internal class ObjectQuery<E : BaseRealmObject> constructor(
             mediator,
             classKey,
             clazz,
-            property,
+            classMetadata!!.getOrThrow(property),
             type,
             AggregatorQueryType.MIN
         )
@@ -160,13 +157,21 @@ internal class ObjectQuery<E : BaseRealmObject> constructor(
             mediator,
             classKey,
             clazz,
-            property,
+            classMetadata!!.getOrThrow(property),
             type,
             AggregatorQueryType.MAX
         )
 
     override fun <T : Any> sum(property: String, type: KClass<T>): RealmScalarQuery<T> =
-        SumQuery(realmReference, queryPointer, mediator, classKey, clazz, property, type)
+        SumQuery(
+            realmReference,
+            queryPointer,
+            mediator,
+            classKey,
+            clazz,
+            classMetadata!!.getOrThrow(property),
+            type
+        )
 
     override fun count(): RealmScalarQuery<Long> =
         CountQuery(realmReference, queryPointer, mediator, classKey, clazz)
@@ -192,14 +197,14 @@ internal class ObjectQuery<E : BaseRealmObject> constructor(
 
     companion object {
         private fun parseQuery(
-            dbPointer: RealmPointer,
+            realmReference: RealmReference,
             classKey: ClassKey,
             filter: String,
             args: Array<out Any?>
         ): RealmQueryPointer = tryCatchCoreException {
             inputScope {
                 val queryArgs = convertToQueryArgs(args)
-                RealmInterop.realm_query_parse(dbPointer, classKey, filter, queryArgs)
+                RealmInterop.realm_query_parse(realmReference.dbPointer, classKey, filter, queryArgs)
             }
         }