+++ packages/test-sync/src/commonMain/kotlin/io/realm/kotlin/entities/sync/SyncObjectWithAllTypes.kt
@@ -16,11 +16,13 @@
 
 package io.realm.kotlin.entities.sync
 
+import io.realm.kotlin.ext.asRealmObject
 import io.realm.kotlin.ext.realmListOf
 import io.realm.kotlin.ext.realmSetOf
 import io.realm.kotlin.schema.RealmStorageType
 import io.realm.kotlin.types.MutableRealmInt
 import io.realm.kotlin.types.ObjectId
+import io.realm.kotlin.types.RealmAny
 import io.realm.kotlin.types.RealmInstant
 import io.realm.kotlin.types.RealmList
 import io.realm.kotlin.types.RealmObject
@@ -71,6 +73,8 @@ class SyncObjectWithAllTypes : RealmObject {
     var binaryNullableField: ByteArray? = null
     var objectNullableField: SyncObjectWithAllTypes? = null
     var mutableRealmIntNullableField: MutableRealmInt? = null
+    var nullableRealmAnyField: RealmAny? = null
+    var nullableRealmAnyForObjectField: RealmAny? = null
 
     // RealmLists
     var stringRealmList: RealmList<String> = realmListOf("hello world")
@@ -87,6 +91,8 @@ class SyncObjectWithAllTypes : RealmObject {
     var realmUUIDRealmList: RealmList<RealmUUID> = realmListOf(RealmUUID.random())
     var binaryRealmList: RealmList<ByteArray> = realmListOf(byteArrayOf(42))
     var objectRealmList: RealmList<SyncObjectWithAllTypes> = realmListOf()
+
+    var nullableRealmAnyRealmList: RealmList<RealmAny?> = realmListOf(RealmAny.create(42))
 
     // Nullable RealmLists of primitive values, not currently supported by Sync
     // Nullable Object lists, not currently supported by Core
@@ -106,6 +112,8 @@ class SyncObjectWithAllTypes : RealmObject {
     var realmUUIDRealmSet: RealmSet<RealmUUID> = realmSetOf(RealmUUID.random())
     var binaryRealmSet: RealmSet<ByteArray> = realmSetOf(byteArrayOf(42))
     var objectRealmSet: RealmSet<SyncObjectWithAllTypes> = realmSetOf()
+
+    var nullableRealmAnyRealmSet: RealmSet<RealmAny?> = realmSetOf(RealmAny.create(42))
 
     // RealmSets of nullable primitive values, not currently supported by Sync
     // RealmSets of nullable objects, not currently supported by Core
@@ -373,6 +381,78 @@ class SyncObjectWithAllTypes : RealmObject {
                                             byteArrayOf(11, 33),
                                             obj.binaryRealmSet
                                         )
+                                    },
+                                )
+                            }
+                            RealmStorageType.ANY -> {
+                                val realmAnyValues = listOf(
+                                    RealmAny.create(42),
+                                    RealmAny.create("hello"),
+                                    RealmAny.create(
+                                        SyncObjectWithAllTypes().apply {
+                                            stringField = "Custom"
+                                        }
+                                    )
+                                )
+                                // Don't reuse the same object in collections as we would be saving the same PK
+                                val realmAnyListValues = listOf(
+                                    realmAnyValues[0],
+                                    realmAnyValues[1],
+                                    RealmAny.create(
+                                        SyncObjectWithAllTypes().apply {
+                                            stringField = "List_element"
+                                        }
+                                    )
+                                )
+                                val realmAnySetValues = listOf(
+                                    realmAnyValues[0],
+                                    realmAnyValues[1],
+                                    RealmAny.create(
+                                        SyncObjectWithAllTypes().apply {
+                                            stringField = "Set_element"
+                                        }
+                                    )
+                                )
+                                Pair(
+                                    { obj: SyncObjectWithAllTypes ->
+                                        obj.nullableRealmAnyField = realmAnyValues[0]
+                                        obj.nullableRealmAnyForObjectField = realmAnyValues[2]
+                                        obj.nullableRealmAnyRealmList = realmListOf(realmAnyListValues[0], realmAnyListValues[1], realmAnyListValues[2], null)
+                                        obj.nullableRealmAnyRealmSet = realmSetOf(realmAnySetValues[0], realmAnySetValues[1], realmAnySetValues[2], null)
+                                    },
+                                    { obj: SyncObjectWithAllTypes ->
+                                        // Check RealmAny containing an object
+                                        assertEquals(
+                                            realmAnyValues[2].asRealmObject<SyncObjectWithAllTypes>().stringField,
+                                            obj.nullableRealmAnyForObjectField?.asRealmObject<SyncObjectWithAllTypes>()?.stringField
+                                        )
+
+                                        // Check RealmAny field containing a primitive
+                                        assertEquals(realmAnyValues[0], obj.nullableRealmAnyField)
+
+                                        // Check list of RealmAny values
+                                        assertEquals(realmAnyListValues[0], obj.nullableRealmAnyRealmList[0])
+                                        assertEquals(realmAnyListValues[1], obj.nullableRealmAnyRealmList[1])
+                                        assertEquals(
+                                            realmAnyListValues[2].asRealmObject<SyncObjectWithAllTypes>().stringField,
+                                            obj.nullableRealmAnyRealmList[2]?.asRealmObject<SyncObjectWithAllTypes>()?.stringField
+                                        )
+                                        assertEquals(null, obj.nullableRealmAnyRealmList[3])
+
+                                        // Check set of RealmAny values
+                                        assertSetContains(realmAnySetValues[0], obj.nullableRealmAnyRealmSet)
+                                        assertSetContains(realmAnySetValues[1], obj.nullableRealmAnyRealmSet)
+                                        assertSetContains(null, obj.nullableRealmAnyRealmSet)
+
+                                        // Extremely irritating to check this since none of the helpers are useful
+                                        obj.nullableRealmAnyRealmSet
+                                            .first {
+                                                it?.type == RealmAny.Type.OBJECT
+                                            }.also {
+                                                val expected = realmAnySetValues[2].asRealmObject<SyncObjectWithAllTypes>().stringField
+                                                val actual = it?.asRealmObject<SyncObjectWithAllTypes>()?.stringField
+                                                assertEquals(expected, actual)
+                                            }
                                     },
                                 )
                             }